requirements.txt
better_profanity==0.7.0
Django==5.0.7
django_ckeditor==6.7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
geopy==2.4.1
Pillow==10.4.0
protobuf==5.27.2
PyJWT==2.8.0
PyJWT==2.8.0
pytz==2024.1
Requests==2.32.3
Shapely==2.0.5
django-cors-headers
google-generativeai
mysqlclient

